% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_booklog.R
\name{read_booklog}
\alias{read_booklog}
\title{Read booklog.csv into a Data Frame}
\usage{
read_booklog(file = "booklog.csv", ...)
}
\arguments{
\item{file}{A character string containing the file to be read in, or a file connection. Defaults to booklog.csv.}

\item{...}{Further arguments to be passed to read.csv.}
}
\value{
Returns a data frame with the contents of booklog.csv.
}
\description{
This is just a wrapper around read.csv to read booklog.csv into a data frame.
}
\examples{
blg <- read_booklog()
}

